Move components dir
