+++ js/components/HomeScene/Models/NorthfaceModel.jsx
@@ -7,7 +7,7 @@ import { useGLTF, useAnimations } from "@react-three/drei";
 import * as THREE from "three";
 
 import { GLB_ASSET_URLS } from "@/js/core/constants";
-import { Box } from "../Scene/Common";
+import { Box } from "../Common";
 
 export default function NorthfaceModel(props) {
     const [mounted, setMounted] = useState(false);
